requirements/base.txt
Django~=5.0
django-environ==0.11.2
psycopg==3.1.14
djangorestframework==3.15.2
dj-rest-auth==6.0.0
